[MOV] core: move Query into odoo.orm
It is already linked to the ORM and will be further integrated.
