+++ addons/purchase/report/purchase_report.py
@@ -1,12 +1,11 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 #
 # Please note that these reports are not multi-currency !!!
 #
 
-from odoo import fields, models, api
-from odoo.tools.query import Query
+from odoo import fields, models
+from odoo.models import Query
 from odoo.tools.sql import SQL
