+++ odoo/tools/__init__.py
@@ -12,7 +12,6 @@ from .i18n import format_list, py_to_js_locale
 from .json import json_default
 from .mail import *
 from .misc import *
-from .query import Query
 from .sql import *
 from .translate import _, html_translate, xml_translate, LazyTranslate
 from .xml_utils import cleanup_xml_node, load_xsd_files_from_url, validate_xml_from_attachment
